users/views.py
# users/views.py
from rest_framework import viewsets
from .models import Users, Role, Permission
from .serializer import UsersSerializer, RoleSerializer, PermissionSerializer

class UsersView(viewsets.ModelViewSet):
    queryset = Users.objects.all()
    serializer_class = UsersSerializer

class RoleView(viewsets.ModelViewSet):
    queryset = Role.objects.all()
    serializer_class = RoleSerializer

class PermissionView(viewsets.ModelViewSet):
    queryset = Permission.objects.all()
    serializer_class = PermissionSerializer